+++ CHANGELOG.md
@@ -1,8 +1,13 @@
 # Changelog
+
+## 0.3.0
+
+* `sources` and `layers` are now in the `cartoj.core` namespace
+* The view helpers were moved out of core into `cartoj.interop`
 ## 0.2.0
 
 - Make the `pmtiles` npm library optional. #1
 
 ## 0.1.0
 
-Initial Release
+* Initial Release